+++ apps/client/src/app/pages/home/home-page.component.ts
@@ -1,8 +1,6 @@
 import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
-import { DataService } from '@ghostfolio/client/services/data.service';
 import { UserService } from '@ghostfolio/client/services/user/user.service';
 import { TabConfiguration, User } from '@ghostfolio/common/interfaces';
-import { hasPermission, permissions } from '@ghostfolio/common/permissions';
 import { DeviceDetectorService } from 'ngx-device-detector';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
@@ -15,7 +13,6 @@ import { takeUntil } from 'rxjs/operators';
 })
 export class HomePageComponent implements OnDestroy, OnInit {
   public deviceType: string;
-  public hasPermissionToAccessFearAndGreedIndex: boolean;
   public tabs: TabConfiguration[] = [];
   public user: User;
 
@@ -23,17 +20,9 @@ export class HomePageComponent implements OnDestroy, OnInit {
 
   public constructor(
     private changeDetectorRef: ChangeDetectorRef,
-    private dataService: DataService,
     private deviceService: DeviceDetectorService,
     private userService: UserService
   ) {
-    const { globalPermissions } = this.dataService.fetchInfo();
-
-    this.hasPermissionToAccessFearAndGreedIndex = hasPermission(
-      globalPermissions,
-      permissions.enableFearAndGreedIndex
-    );
-
     this.userService.stateChanged
       .pipe(takeUntil(this.unsubscribeSubject))
       .subscribe((state) => {
@@ -57,8 +46,7 @@ export class HomePageComponent implements OnDestroy, OnInit {
             {
               iconName: 'newspaper-outline',
               label: $localize`Markets`,
-              path: ['/home', 'market'],
-              showCondition: this.hasPermissionToAccessFearAndGreedIndex
+              path: ['/home', 'market']
             }
           ];
           this.user = state.user;