+++ apps/client/src/app/pages/landing/landing-page.module.ts
@@ -4,6 +4,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { RouterModule } from '@angular/router';
 import { GfWorldMapChartModule } from '@ghostfolio/client/components/world-map-chart/world-map-chart.module';
+import { GfCarouselModule } from '@ghostfolio/ui/carousel';
 import { GfLogoModule } from '@ghostfolio/ui/logo';
 import { GfValueModule } from '@ghostfolio/ui/value';
 
@@ -14,6 +15,7 @@ import { LandingPageComponent } from './landing-page.component';
   declarations: [LandingPageComponent],
   imports: [
     CommonModule,
+    GfCarouselModule,
     GfLogoModule,
     GfValueModule,
     GfWorldMapChartModule,